pyproject.toml
[tool.poetry]
name = "fastapi-manga"
version = "0.1.0"
description = ""
authors = ["Doctor <thirvondukr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
aiofiles = "^0.7.0"
uvicorn = "^0.15.0"
asyncpg = "^0.24.0"
websockets = "^9.1"
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.6.5"
python-slugify = "^5.0.2"
pydantic = {extras = ["email", "dotenv"], version = "^1.8.2"}
strawberry-graphql = "^0.78.0"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.25"}

[tool.poetry.dev-dependencies]
black = "^21.7b0"
Faker = "^8.12.1"
opentelemetry-api = "^1.5.0"
opentelemetry-sdk = "^1.5.0"
tox = "^3.24.4"
mypy = "^0.910"
types-python-slugify = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
